+++ .rebase/CHANGELOG.md
@@ -3,6 +3,22 @@
 The file to keep a list of changed files which will potentionaly help to resolve rebase conflicts.
 
 #### @RomanNikitenko
+https://github.com/che-incubator/che-code/pull/655
+
+- code/src/vs/platform/shell/node/shellEnv.ts
+- code/src/vs/server/node/remoteTerminalChannel.ts
+- code/src/vs/server/node/extensionHostConnection.ts
+---
+
+#### @sbouchet
+https://github.com/che-incubator/che-code/pull/647
+
+- code/test/monaco/package.json
+- code/build/package.json
+- code/package.json
+---
+
+#### @RomanNikitenko
 https://github.com/che-incubator/che-code/pull/617
 
 - code/src/vs/platform/extensionManagement/common/extensionManagement.ts